integrated async/await to some routes

+++ server/server.js
@@ -67,23 +67,21 @@ app.get( '/todos/:id', authenticate, ( req, res ) =>
 	});
 });
 
-app.delete( '/todos/:id', authenticate, ( req, res ) =>
+app.delete( '/todos/:id', authenticate, async ( req, res ) =>
 {
-	var id = req.params.id;
+	let id = req.params.id;
 
 	if ( ! ObjectID.isValid( id ) ) return res.status( 404 ).send( 'Invalid ID!' );
-
-	Todo.findOneAndRemove(
+	
+	try
 	{
-		_id: id,
-		_creator: req.user._id
-	}).then( ( todo ) =>
-	{
+		let todo = await Todo.findOneAndRemove( { _id: id, _creator: req.user._id } );
 		todo ? res.send( { todo } ) : res.status( 404 ).send( 'Todo does not exist!' );
-	}, ( e ) =>
+	}
+	catch ( e )
 	{
 		res.status( 400 ).send();
-	});
+	}
 });
 
 app.patch( '/todos/:id', authenticate, ( req, res ) =>
@@ -113,22 +111,20 @@ app.patch( '/todos/:id', authenticate, ( req, res ) =>
 	});
 });
 
-app.post( '/users', ( req, res ) =>
+app.post( '/users', async ( req, res ) =>
 {
-	var body = _.pick( req.body, [ 'email', 'password' ] );
-
-	var user = new User( body );
-
-	user.save().then( () =>
+	try
 	{
-		return user.generateAuthToken();
-	}).then( ( token ) =>
-	{
+		let body = _.pick( req.body, [ 'email', 'password' ] );
+		let user = new User( body );
+		await user.save();
+		let token = await user.generateAuthToken();
 		res.header( 'x-auth', token ).send( user );
-	}).catch( ( e ) =>
+	}
+	catch ( e )
 	{
 		res.status( 400 ).send( e );
-	});
+	}
 });
 
 app.get( '/users/me', authenticate, ( req, res ) =>
@@ -136,28 +132,32 @@ app.get( '/users/me', authenticate, ( req, res ) =>
 	res.send( req.user );
 });
 
-app.post( '/users/login', ( req, res ) =>
+app.post( '/users/login', async ( req, res ) =>
 {
-	var body = _.pick( req.body, [ 'email', 'password' ] );
-
-	User.findByCredentials( body.email, body.password ).then( ( user ) =>
+	try
 	{
-		return user.generateAuthToken().then( ( token ) =>
-		{
-			res.header( 'x-auth', token ).send( user );
-		});
-	}).catch( ( e ) => res.status( 400 ).send() );
+		let body = _.pick( req.body, [ 'email', 'password' ] );
+		let user = await User.findByCredentials( body.email, body.password );
+		let token = await user.generateAuthToken();
+		res.header( 'x-auth', token ).send( user );
+	}
+	catch ( e )
+	{
+		res.status( 400 ).send();
+	}
 });
 
-app.delete( '/users/me/token', authenticate, ( req, res ) =>
+app.delete( '/users/me/token', authenticate, async ( req, res ) =>
 {
-	req.user.removeToken( req.token ).then( () =>
-		{
-			res.status( 200 ).send();
-		}, () =>
-		{
-			res.status( 400 ).send();
-		});
+	try
+	{
+		await req.user.removeToken( req. token );
+		res.status( 200 ).send();
+	}
+	catch ( e )
+	{
+		res.status( 400 ).send();
+	}
 });
 
 app.listen( port, () => console.log( `Started on port ${ port }` ) );
